refactor(ui): consolidate the two SSE-subscription implementations into one
Two competing registry-event SSE subscriptions existed: use-registry-events.ts
(wired in health/index/status, with a mainnet-only guard, re-subscribe on
network/API-base change, and connect-replay skip) and the never-mounted
lib/metagraphed/sse.ts + sse-context.tsx (exposed connection status but lacked
network reactivity, testnet safety, and replay-skip).

Keep the wired, safer hook as the single implementation, fold the one genuinely
missing piece — connection status (status + lastEventAt) — into its return, and
delete the dead sse.ts + sse-context.tsx entirely.

- One EventSource-based file remains (use-registry-events.ts).
- The survivor retains reconnect-on-network-change, the mainnet-only guard, and
  connect-replay skip; it now also returns { status, lastEventAt } for optional
  consumers (e.g. a liveness indicator).
- Deleted the unmounted loser and its context; no dangling imports remain.

Pure hook/lib refactor with no visual change — the three route consumers call the
hook for its side effect and ignore the return.

Closes #3436
